server/db/driver/pg: version the scheme
This adds a schema_version column to the meta table, possibly
creating the table if it was missing. This is the first defined
upgrade to v1.

This also adds a v2 upgrade that partially populates the epoch_stats
table with historical data from the matches table. v2 upgrade creates
epochs_report table, if it does not exist, and populates the table with
partial historical data from the epochs and matches table. This includes
match volumes, high/low/start/end rates, but does not include the booked
volume statistics in the book_buys* and book_sells* columns since this
data requires a book snapshot at the time of matching to generate.

Test upgradeDB from v0 with DB snapshots from the following cases:

- v0 DB with no meta table, as on 0.2+pre master presently
- v0 DB with a meta table with state_hash from release-0.1
- v0 with meta table and a ton of matches for v2 upgrade

The db snapshot archive with matches is quite large
(server/db/driver/pg/dcrdex_test_db_v0-release-0.1-matches.sql.gz),
but it is a good test as it has numberous epochs and matches, pseudo-
randomly generated with loadbot running on release-0.1.