multi: graceful support for json unmarshal of null
A number of json.Unmarshal calls now properly handle a Payload
of the JSON null value, which depending on the caller's use of
json.Unmarshal can set a pointer of the type to nil.

The docs for (*Message).Unmarshal now call out the null Payload
handling explicitly. A null Payload is not an error however, since
the caller may be happy to accept (or even expect) a null Payload.
Further, for the caller to allow their input set to nil, they must
call Unmarshal with a specific type. For example,

  t := new(T) // *T
  err := json.Unmarshal(msg.Payload, &t) // passed **T, can set t to nil

Only with the above approach can json.Unmarshal or (*Message).Unmarshal
set t to nil. Callers can either pass a *T instead and accept a
zero-value T in the case of a null Payload, or pass a **T and check
`if err != nil || t == nil` when they need to identify this case.

For example, the client's RPC server's handleRequest method permits,
and in the case of the version RPC expects, a null Payload without
error and as such unmarshals with:

  params := new(RawParams)
  err := req.Unmarshal(params) // NOT &params - cannot be set to nil for []byte("null") Payload

For the client RPC server, this is the safest approach, and a non-nil
error is required. Similarly, dex/ws.(*WSLink).inhandler expects no
error for a null Payload and recognizes it with a specific check:

  msg := new(msgjson.Message)
  err = json.Unmarshal(msgBytes, msg) // *Message
  ...
  if msg.ID == 0 { // also covers msgBytes []byte("null")
      c.SendError(1, msgjson.NewError(msgjson.RPCParseError, "request id cannot be zero"))

The (*Message).Response method unmarshals with a **T, and is update
with this nil check, and returns an error:
  fmt.Errorf("null response payload").

The docs for msgjson.DecodeMessage are updated to reflect that the
both returned *Message and error can be nil in this null case.
DecodeMessage is only used in tests presently.

Each of the following route handlers require detecting a null payload,
and are thus updated with post Unmarshal checks: handleConnect,
handleRegister, handleNotifyFee, handleInit, and handleRedeem.
All routes were called from a helper with a deferred recover.

This also adds the method client/comms.(*wsConn).SendRaw([]byte) for
testing purposes. It is not part of the WsConn interface however.

swapper: log fix: responses have no route set

test and use (*Message).Unmarshal and Response methods

Also reject null payloads with the orderbook and unsub_orderbook
unauthenticated routes.
Also reject null payloads with hte limit, market, and cancel routes
handled by the OrderRouter since the zero-values are not sensible.